% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cell.R
\name{predict_cell}
\alias{predict_cell}
\title{Cell types prediction.}
\usage{
predict_cell(path_score)
}
\arguments{
\item{path_score}{The pathway score matrix calculated by the pathway_score function.}
}
\value{
A data.frame containing cell types and predicted values.
}
\description{
Classify tumor cells from normal cells.
}
\details{
Input the pathway score matrix calculated by the pathway_score function.
}
