% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalized.R
\name{data_normalized}
\alias{data_normalized}
\title{data normalized}
\usage{
data_normalized(expr_data)
}
\arguments{
\item{expr_data}{A single-cell counts expression matrix.}
}
\value{
A normalized single-cell expression matrix.
}
\description{
Normalize single-cell raw counts matrix.
}
\details{
Input a data.frame where the rows are the gene names and the columns are the sample names.
}
