% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_score.R
\name{pathway_score}
\alias{pathway_score}
\title{pathway score}
\usage{
pathway_score(normalized_matrix)
}
\arguments{
\item{normalized_matrix}{Single-cell expression matrix after normalization of the original counts matrix.}
}
\value{
A matrix containing 213 pathway scores.
}
\description{
Obtain a pathway score matrix for predicting tumor cells.
}
\details{
Input a data.frame where the rows are the gene names and the columns are the sample names. Matrix that can be generated directly using the data_normalized.R function.
}
